HubSpot Backport: HBASE-29131 Introduce the option for post-compaction validation of HFiles (#6700) (will be in 2.6.3)
Introduces the option for an HStore to fully read the file it just wrote after a flush or
compaction.

To enable this feature, set `hbase.hstore.validate.read_fully=true`. This is an HStore
configuration feature, so it can be enabled in hbase-site.xml, in the TableDescriptor, or in the
ColumnFamilyDescriptor.

Signed-off-by: Peter Somogyi <psomogyi@apache.org >