Clear meta cache for whole server on certain exceptions
When a server hosting many regions is replaced, clients spend many
requests clearing it from the meta cache one region at a time as they
encounter "Connection refused" ConnectExceptions. If they don't clear
all regions in time before the host OS goes offline (and stops
fast-rejecting TCP connections), subsequent requests will need to wait
for an entire CallTimeout duration before they can succeed or be
retried.

Instead, clear the meta cache for a server when we encounter exceptions
which indicate the server is not hosting any regions. The larger this
exception set, the more aggressively clients will make requests to the
meta region, so this change conservatively omits timeout exceptions,
although it could be made configurable in the future.